Extract FeedImageDataLoader test helpers into a shared protocol ext…
…ension to remove duplication
